Changed folder structure for create tools
