+++ backend/app/routes/stages/create/ai_art_creator_for_kids_auth.py
