+++ backend/app/routes/stages/create/interactive_comic_strip_builder_auth.py
