+++ backend/app/routes/stages/create/learn_by_drawing_auth.py
