+++ backend/app/routes/stages/create/three_d_model_builder_auth.py
