+++ backend/app/services/stages/create/ai_art_creator_for_kids_service.py
