+++ backend/app/services/stages/create/learn_by_drawing_service.py
