+++ backend/app/services/stages/create/three_d_model_builder_service.py
