+++ frontend/src/stages/Create/Create.js
@@ -19,10 +19,10 @@ const createTools = [
   { name: 'Creative Writing Prompts', path: '/creative_writing_prompts', icon: <FaLightbulb />, colorClass: 'card-purple' },
   { name: 'Story Visualizer', path: '/story_visualizer', icon: <FaImages />, colorClass: 'card-pink' },
   { name: 'Story to Comics Converter', path: '/story_to_comics', icon: <FaFilm />, colorClass: 'card-teal' },
-  { name: 'Interactive Comic Strip Builder', path: '/comic_strip_builder', icon: <FaProjectDiagram />, colorClass: 'card-green' },
+  { name: 'Interactive Comic Strip Builder', path: '/interactive_comic_strip_builder', icon: <FaProjectDiagram />, colorClass: 'card-green' },
   { name: 'AI Presentation Builder', path: '/ai_presentation_builder', icon: <FaChalkboardTeacher />, colorClass: 'card-amber' },
-  { name: 'AI Art Creator for Kids', path: '/ai_art_creator_kids', icon: <FaPaintBrush />, colorClass: 'card-blue' },
-  { name: '3D Model Builder', path: '/model_builder_3d', icon: <FaCube />, colorClass: 'card-teal' },
+  { name: 'AI Art Creator for Kids', path: '/ai_art_creator_for_kids', icon: <FaPaintBrush />, colorClass: 'card-blue' },
+  { name: '3D Model Builder', path: '/three_d_model_builder', icon: <FaCube />, colorClass: 'card-teal' },
   { name: 'Data Story Builder', path: '/data_story_builder', icon: <FaChartArea />, colorClass: 'card-amber' },
   { name: 'Learn by Drawing', path: '/learn_by_drawing', icon: <FaPencilRuler />, colorClass: 'card-green' },
   { name: 'Historical Timeline Builder', path: '/historical_timeline_builder', icon: <FaHistory />, colorClass: 'card-purple' },
