+++ frontend/src/stages/Create/three-d-model-builder.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
-import './story-to-comics-converter.css';
-import config from '../config';
+import './three-d-model-builder.css';
+import config from '../../config';
 
-export default function Story_to_comics_converterTool({ fileId = null }) {
+export default function Three_d_model_builderTool({ fileId = null }) {
   const [progressId, setProgressId] = useState(null);
   const [progress, setProgress] = useState({ percentage: 0, status: 'idle' });
   const [result, setResult] = useState(null);
@@ -15,7 +15,7 @@ export default function Story_to_comics_converterTool({ fileId = null }) {
     setError('');
     setResult(null);
     try {
-      const resp = await fetch(`${config.API_BASE_URL}/story-to-comics-converter/start`, {
+      const resp = await fetch(`${config.API_BASE_URL}/3d-model-builder/start`, {
         method: 'POST',
         headers: auth,
         body: JSON.stringify({ file_id: fileId, params: {} })
@@ -30,13 +30,13 @@ export default function Story_to_comics_converterTool({ fileId = null }) {
   useEffect(() => {
     if (!progressId) return;
     let timer = setInterval(async () => {
-      const r = await fetch(`${config.API_BASE_URL}/story-to-comics-converter/progress/${progressId}`, { headers: auth });
+      const r = await fetch(`${config.API_BASE_URL}/3d-model-builder/progress/${progressId}`, { headers: auth });
       const d = await r.json();
       if (!r.ok) { setError(d.error || 'Progress error'); clearInterval(timer); return; }
       setProgress({ percentage: d.percentage || 0, status: d.status });
       if (d.status === 'done') {
         clearInterval(timer);
-        const rr = await fetch(`${config.API_BASE_URL}/story-to-comics-converter/results/${progressId}`, { headers: auth });
+        const rr = await fetch(`${config.API_BASE_URL}/3d-model-builder/results/${progressId}`, { headers: auth });
         const rd = await rr.json();
         if (rr.ok) setResult(rd); else setError(rd.error || 'Results error');
       }
@@ -46,9 +46,9 @@ export default function Story_to_comics_converterTool({ fileId = null }) {
   }, [progressId]);
 
   return (
-    <div className="story-to-comics-converter-tool">
+    <div className="three-d-model-builder-tool">
       <div className="tool-head">
-        <h3>Story to Comics Converter</h3>
+        <h3>3D Model Builder</h3>
         <button className="btn" onClick={start}>Run</button>
       </div>
